# docs/swagger.yaml
basePath: /
definitions:
...
  erg_ninja_internal_modules_elearning_dto.CategoryResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      levels:
        items:
          $ref: '#/definitions/erg_ninja_internal_modules_elearning_dto.LevelResponse'
        type: array
      name:
        type: string
      order:
        type: integer
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  erg_ninja_internal_modules_elearning_dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      order:
        type: integer
      parent_id:
        type: string
      slug:
        type: string
      tenant_id:
        type: string
    required:
    - name
    type: object
  erg_ninja_internal_modules_elearning_dto.LevelResponse:
    properties:
      category_id:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      order:
        type: integer
      slug:
        type: string
      units:
        items:
          $ref: '#/definitions/erg_ninja_internal_modules_elearning_dto.UnitResponse'
        type: array
      updatedAt:
        type: string
    type: object
  erg_ninja_internal_modules_elearning_dto.UnitResponse:
    properties:
      content:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      level_id:
        type: string
      name:
        type: string
      order:
        type: integer
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: ERG Support
    url: https://erg.ninja/support
  description: The unified backend API for ERG — a multi-tenant eLearning, content
    management, and automation platform.
  title: ERG Platform API
  version: "1.0"
paths:
  /admin/elearning/categories:
    get:
      consumes:
      - application/json
      description: Fetch all categories for admin management.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/erg_ninja_internal_modules_elearning_dto.CategoryResponse'
            type: array
      security:
      - BearerAuth: []
      summary: 'Admin: List categories'
      tags:
      - Elearning Admin
    post:
      consumes:
      - application/json
      description: Create a new category.
      parameters:
      - description: Category Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/erg_ninja_internal_modules_elearning_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/erg_ninja_internal_modules_elearning_dto.CategoryResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Create category'
      tags:
      - Elearning Admin
  /admin/elearning/categories/{id}:
    delete:
      description: Deletes an elearning category.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: Delete category'
      tags:
      - Elearning Admin
    patch:
      consumes:
      - application/json
      description: Updates an elearning category.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: Update category'
      tags:
      - Elearning Admin
  /admin/elearning/levels:
    post:
      consumes:
      - application/json
      description: Creates a new elearning level.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: Create level'
      tags:
      - Elearning Admin
  /admin/elearning/levels/{id}:
    delete:
      description: Deletes an elearning level.
      parameters:
      - description: Level ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: Delete level'
      tags:
      - Elearning Admin
    patch:
      consumes:
      - application/json
      description: Updates an elearning level.
      parameters:
      - description: Level ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: Update level'
      tags:
      - Elearning Admin
  /admin/elearning/units:
    post:
      consumes:
      - application/json
      description: Creates a new elearning unit.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: Create unit'
      tags:
      - Elearning Admin
  /admin/elearning/units/{id}:
    delete:
      description: Deletes an elearning unit.
      parameters:
      - description: Unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: Delete unit'
      tags:
      - Elearning Admin
    patch:
      consumes:
      - application/json
      description: Updates an elearning unit.
      parameters:
      - description: Unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: Update unit'
      tags:
      - Elearning Admin
...
  /api/elearning/categories:
    get:
      consumes:
      - application/json
      description: Fetch all categories for the current tenant.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/erg_ninja_internal_modules_elearning_dto.CategoryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List categories
      tags:
      - Elearning
  /api/elearning/categories/{slug}:
    get:
      consumes:
      - application/json
      description: Fetch category details by slug.
      parameters:
      - description: Category Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erg_ninja_internal_modules_elearning_dto.CategoryResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get category by slug
      tags:
      - Elearning
  /api/elearning/levels/{slug}:
    get:
      consumes:
      - application/json
      description: Fetch level details by slug.
      parameters:
      - description: Level Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erg_ninja_internal_modules_elearning_dto.LevelResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get level by slug
      tags:
      - Elearning
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"